Convert to Elasticsearch streaming_bulk helper
This creates a new ElasticBulkBase class providing support for
Elasticsearch bulk operations including publish and delete, and
which currently shares the query_apis `__init__.py` with the
ElasticBase class.

This commit "completes" the refactoring, however I haven't
built unit tests yet, and more functional testing is required to
take it out of draft.