Use generator pattern w multi-processing queues
* Create `load_pbench_runs()` method

The pbench run data is much smaller than the result data, so we
create a very simple routine to load all the run data, apply the
proper filters, massage it a bit, and store only the fields we
need when processing the result data.  This provides a much smaller
memory footprint.

* Use a generator pattern for processing results

We use a generator to load one record at a time from the JSON
docs loaded from a file, process it, adorn it with pbench run
data and then write it out without storing it.  This avoids an
using memory to store all the data while processing.

This does not implement the client names and/or disk/host names
fetching required to complete the data adornment.