Pull from Elasticsearch directly
Instead of using separate `curl` commands to pull the data from
Elasticsearch we just use Python3 to do it directly.  This reduces
the memory usage substantially, and appears to speed up the time
it takes to process all the data as well.

We also drop code no longer used, leaving a few routines to be
folded in later.