More cleanup
Reduced the bulk failure logging to a "report" document showing
counts for each error reason in each index, which might be a bit
more useful than just error counts but is much less bulky than the
details of each error. The return value to client is a new "summary"
document with just ok/failure counts.