Cleanup in JSON parameter handling
While attempting to debug a Postman problem I tried removing
the "silent=True" option to avoid hiding trouble. (With this
option, `get_json` simply returns `None`.) This exposed a
problem in `GET` handling, in that we were still attempting to
process a JSON payload when there is none and we won't
use it.

I added a `try`/`except` to catch and report any errors here, and
to abort with BAD_REQUEST. I also added a new unit test with
deliberately malformed JSON.

None of this unfortunately helped to make progress on the original
problem, but it seems like a useful cleanup.