Add variable for the pbench repo name
Fixes #2288.

Add an ansible variable to the defaults so that the pbench repo can be
customized by the user. The variable is named `pbench_repo_name` and
its default value is `pbench`. Together with the existing variable
`fedoraproject_username` and its default value, the two specify the
standard pbench repo.  But just as `fedoraproject_username` can be
changed in the inventory file to point to a user-specific CORP repo,
the `pbench_repo_name` can now be changed to point to a different repo
under that user, e.g. `pbench-test` which is commonly used for test
builds.

Do this for both the agent and server `pbench-repo-install` roles.