Add support for vsock to the pbench-uperf script
vsock protocol does not share the same hostnames as the usual ip stack,
therefore it's not possible to use the same hostname for ssh and uperf
address. Let's add a "--vsock-servers" option to let users specify the
vsock addresses used by uperf in the xml file. Also set the vsock
protocol in such case.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>
Signed-off-by: Peter Portante <peter.portante@redhat.com>