Replace RPM version checks with `fio --version`
We no longer check for the `fio` RPM version, but the output of
`fio --version`.  We use the expected version as recorded in the
controller's `pbench-agent.cfg` file.  This allows us to reduce
the number of `ssh` commands we need to issue per host to one
"pre-check" operation (replaces `--install` behavior), and one
fio server setup per sample.