Fix `/api/v1/endpoints` in container environment
It may be that we need to ensure that pbench-server.cfg has a
valid `host` value for other reasons, but `/api/v1/endpoints` is
all about making sure an API client (A) knows about all the
endpoints we support and (B) can call them reliably using the
recommended path (e.g., through an NGINX reverse proxy).

We had advertised endpoints (in the absense of reverse proxy
forwarding headers) relative to the static `pbench-server.cfg`
host value; this however is unnecessary as the incoming HTTP
Requests object includes the actual access path, which makes
a much better fallback. This PR changes the endpoints API to
use that value, which has the advantage of working within a
container regardless of the pod virtual networking configuration.