Correct the handling of `localhost` client/server
For `pbench-uperf`, we no longer `ssh` unconditionally to a uperf server.
For both clients and servers we only `ssh` remotely if that client or
server is not `127.0.0.1` or `localhost`.

This allows us to unify behavior checks, and it corrects a bug where the
JSON data for single clients of `localhost` will be recorded as such.