Dataset DELETE
The delete operation builds on the PUBLISH operation, but the bulk Elastic
operation to delete all of the indexed Elastic documents is only the first
stage; we also need to delete the Dataset DB row itself and the on-disk
file structure representing the dataset (ARCHIVE tarball and MD5 file, the
unpacked INCOMING tarball if any, and the RESULTS link).

To accomplish that, this PR adds a new module with a set of related classes
to discover and manipulate the on-disk file structure, including the ability
to create and delete datasets.

The create operation subsumes the operation of pbench-server-prep-shim-002
so that it's no longer needed. (NOTE: this PR doesn't actually remove it,
as the impact on the legacy test system felt like more churn than really
necessary here.) The PUT API has been refactored extensively to utilize
the FileTree to create a dataset, and no longer stores the temporary file
in the pbench-server-prep-shim-002 intake directory.