Refactor `pbench-uperf`
Cleanup `pbench-uperf`: consolidate `shift` use and remove a trailing
semi-colon.

Add missing `usage`, and use `stderr`.

Remove unused `break` statement.

Fix options `case` statement indentation.

Drop use of `sed` for comma replacement.

Correct the handling of `localhost` client/server; for `pbench-uperf`,
we no longer `ssh` unconditionally to a uperf server.  For both clients
and servers we only `ssh` remotely if that client or server is not
`127.0.0.1` or `localhost`.  This allows us to unify behavior checks,
and it corrects a bug where the JSON data for single clients of
`localhost` will be recorded as such.

Remove deprecated `pbench-uperf` options.