pbench-register-tool-set: check for bad remotes file (#2514)
Fixes #2463.

Passing a non-existent remotes files with `--remotes=@somefile` causes
multiple usage messages from `pbench-register-tool`: one per tool in
the set. The problem is that `pbench-register-tool-set` does not check
the remote file at all: it just blindly passes it to
`pbench-register-tool` which does detect the error, but since it is
called multiple times, it detects and reports the error multiple
times. `pbench-register-tool-set` needs to check the file itself and
quit if the file is bad (non-existent or ill-formed).

To that end, we add a `--test-arg` option to `pbench-register-tool`
so that it can be used as a checking tool. `pbench-register-tool-set`
calls it once with that option and if that call fails, we get the
error message(s) from that failed call, `pbench-register-tool-set`
produces its usage message *once* and we are done.

A couple of test cases are added to the unit tests to test the two
cases.

A bunch of cleanups to `pbench-register-tool` were added during review
and some gold files were amended to conform to these changes.