Replace RPM version checks with `fio --version` and `uperf -V`
* Replace RPM version checks with `fio --version`

We no longer check for the `fio` RPM version, but the output of
`fio --version`.  We use the expected version as recorded in the
controller's `pbench-agent.cfg` file.  This allows us to reduce
the number of `ssh` commands we need to issue per host to one
"pre-check" operation (replaces `--install` behavior), and one
fio server setup per sample.

* Rename `pbench-uperf --install` to `--pre-check`

In line with `pbench-fio`, we rename the `--install` option, and refactor
the code to use a single `local_pre_check()` function shared between the
local and remote `--pre-check` operations.

* Run version check of `uperf -V` output only

We no longer check for an RPM being installed, but only that the `uperf`
command that we found emits a supported version.