Revamp how we use tox to reduce envs
In summary, by default tox runs all the tests, while accepting arguments
to request various sub-sets to be run.  See the `README.md` file for a
description of the interface.

We also simplify the `jenkins/run` script to remove the `CONTAINER_HOME`
options and just build out `/home/developer` inside the container. There
is no need for an external mounted `/home`.

The default jenkins job will use the default installed python, RPM
provided python modules, and pip install the rest as normal.  The
container use for the CI jobs is built on Fedora 34, using Python 3.9.

A few notes:

  * We tried to remove `whitelist_externals` but for whatever reason it
    seems the `allowlist_externals` is not really allowed just yet

  * We now use `/usr/bin/time` for all sub-system test invocations so
    that we can see how long they took to run individually

  * detox tox without detox

    We found that `utils/detox` is not actually doing anything
    substantially useful prior to this PR. The reason is that if you
    look at the prevoius `tox.ini` file, and the CI job's
    `jenkins/run-pytests`, you'll notice that we run the pytests
    specifying files in the source tree. So all tests then run in the
    source tree, creating `__pycache__` directories all over it. While
    there _are_ `__pycache__` directories under the tox virtualenv
    directories, they are not for our code.

    We are now properly installing our source into the virtualenv as a
    python module.

    We now run the tests by the module path reference so that they
    execute out of the installed sources.

    We work to ensure no other files are created in the source tree and
    use just `/var/tmp` instead of both `/tmp` and `/var/tmp` (on some
    Linux installations, `/tmp` can be a small in-memory file system
    which can easily have its free space exhausted).

    As a result, the need for `utils/detox` is removed.