Parallelize the python & legacy unit tests
We add the use of `pytest-xdist`, which allows us to pick the level
of parallelism via the `-n` switch.  By default we use `-n auto` where
it determines the parallelism to use based on the number of CPUs.

Use `parallel` to run all the legacy tests.  Each agent legacy sub-test
is run as a separate parallel job, and then all the server tests are
individually run through GNU `parallel`.

To facilitate the use of `parallel` for the server legacy tests, we
break out the core server test harness out into a separate script. This
helps to isolate the way tests are run from how the parallelism is
achieved.

We also take the opportunity to use `local` in the various `bash`
scripts, and make variable references consistent.

Finally, we address `mkdir -p` usage to remove its use, where possible,
to ensure setup was performed without any hidden, or unexpected,
directory creation, while fixing a problem in the server unit tests
where the `common` directory was created but not used.