Use Bcrypt-Flask instead of Flask-Bcrypt
Messy community project "Flask-Bcrypt", which has an RPM in Fedora and
derivatives, is not supported and is missing a fix to stop using a
deprecated `bcrypt` interface.  A new project was copied by the somebody
in the community to add the proper fix.  This new project is called,
"Bcrypt-Flask", and propery released on PyPI, but there is no RPM for
it.

It still depends on `bcrypt` (and/or, indirectly, the RPM
`python3-bcrypt`).