Dynamic `root_index_name` (#2572)
Replace hard-coded root-index name with parameter in commons.py
root_index_name in the commons.py file is hard coded to "result-data-sample" where it needs to be script-based. For the namespaces_and_rows PR we need result-data-sample as a root_index_name for finding the appropriate indices, while we need run-toc as root_index_name for finding indices related to that.