Remove implicit use of "quarantiner" user (#2591)
The original "light dataset" implementation provided for dynamically
creating datasets. While most of that was removed, it remained in
the bash base utilities `quarantine` function because that could be
called from `pbench-server-prep-shim-002` before a dataset was
created. However, in normal use the "quarantiner" user doesn't
exist anyway.

This removes that code, so that `quarantine` depends on the dataset
already existing. This includes an adjustment to test-5.2 (which is the
only legacy test case exercising this code through pbench-dispatch)
to pre-create the expected datasets. (Using the "quarantiner" user,
so that the gold file doesn't change!)