pbench-copy-result-tb issue #2437
This is a documentation issue, as the pbench-copy-result-tb
neither documents its parameters well, nor diagnoses missing
or inappropriate parameters.

This is ultimately a minor issue as (first off) this is a utility
command normally run indirectly by pbench-move-results or
pbench-copy-results and (secondly) it's specific to 0.69 and
will be retired when we move to a 0.72 production server.

Nevertheless, the behavior can be somewhat improved both by
expanding the --help message and by improving the diagnostic
messages when we find missing parameters or unexpected data
in the temporary tarball staging area.

Resolves #2437