Make the firewall role standalone
Decouple it from the pbench_agent_config role: we want to be able
to use the latter now as well as in the future and we don't need
to punch holes in the firewall now. When it becomes necessary,
we can add the firewall role to the playbook(s), which does not
affect either the github sources (since they contain no playbooks)
or the Galaxy roles.

Bump the Galaxy version to 1.0.5 and rebuild the collection on Galaxy
to eliminate the above interaction.