Remove controller parameter from queries (#2575)
"controller" is becoming just a dataset metadata; perhaps useful to filter based on origin, but not essential. The new /datasets/list PI is based on a SQL query of the Datasets table, supporting filtering by name (substring, e.g., "fio"), controller, start and end time range (which can be used together or independently for a "before" or "after").

The Dataset record now has both `created` and `uploaded` timestamps, where the latter is the old `created` (defaults
to current time of `Dataset` commit) and the new `created` is read from the tarball `metadata.log` file.

Dataset SQL uniqueness is defined now only by the dataset name.