Remove pbench-server-prep-shim-002
PBENCH-474

This removes the pbench-server-prep-shim-002 tool, which no longer serves a
purpose as Server 1.0 will not support scp dataset uploads. This also removes
the "status" pbench-results-host-info.versioned files, which are not used in
the PUT upload process.

This impacts many legacy test gold files, though in mostly trivial and
obvious ways. The exception is test-5.2, which populated the receive directory
with files to test pbench-server-prep-shim-002 behavior on duplicate files,
bad MD5 files, and "prefix" notation; none of which is relevant for Server 1.0
as PUT will leave the bad files on the agent, and we no longer support the
prefix notation.