Remove test parallelism from Jenkins jobs
Not sure exactly what is happening, but the test parallelism we have
in place is not working well all the time.

When upgrading our Jenkins executors from RHEL 7 to RHEL 8.4 & 8.5 we
seem to have run into some problem with the mechanism used for parallelsim
since removing to makes all the tests pass.