Update `push` to change permissions on `quay.io`
PBENCH-703

When we are pushing images to https://quay.io, we ensure that the target
repository is created ahead of time, if it does not already exist, as a
publicly visible repository, with the `core` team is added as an
administrator.

The new `make` target, `repofix`, uses the default distro targets to
ensure all the existing repositories have the appropriate visibility and
permissions.