Add option to allow chroot specs for COPR builds
Add a CHROOTS variable to the agent/rpm/Makefile and an option
to the `copr-cli build' command in rpm.mk, to allow
the specification of chroots to include or exclude in a build.

The default setting for the b0.69 branch excludes all RHEL9 builds:
the 0.69 agent does not work on RHEL9 because of an `scp -r' bug.