Dump all .cmd files in bench-scripts unit tests
This is a forward port of change from the `b0.69` branch, see PR #2090.

In order to make this work correctly we had to also correct the handling
of the mocked `date` and `date_suffix` variables.  This resulted in a
bug being exposed where the `date_suffix` variable should have been
used by `pbench-trafficgen`, and the unit tests `_date_suffix` by the
pbench-fio `test-06`.