Improve Jenkins container run mechanism
There are several small related issues here; one is that we currently
require `--user=root` to run reliable, but we really don't want to be
running our tests as root. This turns out to be mostly required because
of the second issue, and with this change is no longer  required.

The main issue is that the container inherits the HOME environment
variable; for example, in Jenkins, typically /home/jenkins. This value
is used by `python3 setup.py` for `pip install --user` and to install
our package with PBR for testing. Because `/home/jenkins` doesn't exist
in the container, we require `root` to create the path. Instead, we'll
now override the definition of HOME in order to redirect the `--user`
to `/tmp/jenkins` which we can create without privilege.

In conjunction, this removes the default definition of
`EXTRA_PODMAN_SWITCHES` from Pipeline.gy.

There's another "convenience" feature in here useful for local manual
testing using the container from git worktrees; because a worktree is a
mapping of a root branch, we need to map both the worktree directory and
the referenced branch directory inside the container. This uses the
`git rev-parse` command to figure out the environment and add a second
podman `--volume` option if necessary.