Stop using `selinux` virtualenv wrapper
Basically, there is no real `selinux` python module provided via `pip` install.
All that is out there is a "shim" [1], which is really broken.  On module load,
it forks an invocation of the predominate `/usr/bin/python3` program [2] to get
it's list of site packages, and replaces itself with a module load of the
`selinux` module it finds among those site package locations.

The goal of that virtualenv wrapper is to allow a virtualenv using the same
python version as the real installed `selinux` module to work "seamlessly."

But if your virtualenv is for a python version different from what is installed
on the system, then the `selinux` import will fail with an error trying to load
a module that doesn't work with the version of python in the virtualenv.

We no longer use the `selinux` pip installed module.  Instead we provide our
own wrapper which provides pass-through methods for `is_selinux_enabled()` and
`restorecon()` when a `selinux` module is found, and mocks for those same
methods where the `is_selinux_enabled()` method returns `0`, and the
`restorecon()` method raises an exception to help detect logic errors.

For our testing environments, we already provide proper mocks of the `selinux`
module to get the desired behavior (now updated to work with the new wrapper).

[1] https://github.com/pycontribs/selinux
[2] https://github.com/pycontribs/selinux/blob/master/selinux/__init__.py#L76