Remove `BaseCommand` environment manipulations
There was some thinking that the Tool Meister sub-system needed various
environment variables set in the `BaseCommand` class.  Because of how
the `pytest` unit tests are run, where the `os.environ` dictionary is
shared across tests, these variables would show up for any test that
ran subsequent to a test that invoked a `BaseCommand` sub-class.

We experimented with removing them entirely, and could not find any test
that failed as a result.  This also resolves the inadvertant problem
introduced by PR #2606 which caused `tox -e py3-server` runs to fail.

Fixes #2647.