Refactor development container builds
We collapse the two docker files to one, using a Jinja2 template, and remove
all the Red Hat internal references and RPM GPG keys (since those keys are
tied to the loaded repo file).

We are not allowed to publish UBI images where we add Red Hat subscription
bits on top, so we remove the method of publishing the UBI image.