Simplify agent spec file (#2658)
Basic RPM dependencies ensure that python3 and the pip module are always
present when our %post script runs, however it seems that in some cases
the pip3 command is not recognized. Instead, we invoke it directly using
`python3 -m pip`.

This also means that the "pip bootstrap" sequence for RHEL 7 (which
sometimes didn't "take") is no longer useful.