Support non-DNS controller hosts (#2643)
By default, TDS and Redis are started listening on (and remote hosts are
set to connect to) the controller node's `hostname`. If this is not a
routable DNS name, Tool Meister start hangs.

Instead, identify a routable IP address for the host by determining the
origin of ssh connections to the remote hosts. We need to establish that
the remotes are reachable anyway; this allows us to do so early in
startup before creating the local TDS and Redis instances.

If "orchestration" mode isn't selected, or if PBENCH_REDIS_SERVER
(--redis-server) or PBENCH_TOOL_DATA_SINK (--too-data-sink) configuration
options are specified, these values are used regardless, trusting the
external input.