Remove references to logger method of base class
When the server unit tests run in parallel, there is something about how
the Database.base class is modified to have a "logger" that is not
working.

It seems a bad idea in general to do this kind of logging in a database
class, and we should have the logging at the consumers of the database
object.  Given the unit tests are not affected by this change, seems
like it is a good opportunity to make this kind of a change.