Scrub the pytests
Lots of little changes towards a test suite which can be parallelized.

The first change is to remove the use of the session scoped
`pytestconfig`.  Session scoped fixtures make it difficult to share them
across test sub-systems (one has to be aware of how different sub-
systems will interact, which is prone to error).

As a result of dropping `pytestconfig`, `tmp_path` is used instead to
give us a proper temporary directory using a fixture that can handle
the coming parallelism.

In order to share fixtures between the agent unit and functional tests,
a separate on-disk config method allows both fixtures to leverage the
disk setup which is fixed and common.  The environment variable monkey-
patching is moved to non-session scoped fixtures.

A similar discipline is applied to the server-side on-disk configuration
as well.

The `invalid_config` fixture is removed in favor of having the one
test the used to handling the monkeypatching, and then make its sister
fixture, `valid_config`, be marked `autouse`.

We also take the opportunity to remove any unnecessary fixtures.