Correct Tool Meister sub-process termination
If a sub-process that the Tool Meister creates does not terminate
correctly, it can cause problems.  One such example, described in issue
\#2692 [1], was the result of the `subprocess.kill()` method working
but not taking effect immediately, cause another file to be changed
while `tar` was packaging results, leading to failure.

The algorithm used now is:

 1. `subprocess.terminate()`
 2. wait with timeout
 3. `subprocess.kill()`
 4. wait with timeout
 5. close process file descriptors

As part of this work, and to help facilitate the proper closing of
the sub-process file descriptors, we setup read threads for logging
all the stout/stderr data coming from the sub-proceses.

Partial fix for #2691 [2].

[1] https://github.com/distributed-system-analysis/pbench/issues/2692
[2] https://github.com/distributed-system-analysis/pbench/issues/2691