Refactor Tool Meister module, add `Tool` unittests
Refactor the ToolMeister class and the Tool class hierarchy to construct
Tool objects once.  We were running the constructor for each Tool object
once to run the `install()` method, and then again for each "init" and
"start" operation.

Now the Tool object constructor runs once, and all code used to
determine if a tool is installed correctly results in the `install()`
method.

We also take the opportunity to properly assert the expected state each
method of the various Tool classes expect to encounter.

The Tool Meister constructor now validates that the `pbench_install_dir`
and `tmp_dir` arguments are directories that exist, raising an exception
otherwise.

The `.start()` methods for all Tool object sub-classes has been changed
to raise an exception if the `tool_dir` argument is not a real
directory.

Along the way, we also add type hints for all Tool Meister methods and
its constructor, correcting other type hints, and correcting the return
values of a few methods.

Other refactorings:

 * We also add docstrings to the Tool class hierarchy where missing

 * Use named tuples for installation results and tool meister params

 * Refactor {`start`|`init`}`_tools` to share code

 * Refactor the `ToolMeister` class's `start_tools()` and `init_tools()`
   methods to share as much code as possible

 * Rework cleanup of `self._running_tools`

Lastly, we have added unit tests for the Tool class hierarchy and some
helper functions.