Refactor ToolMeister to construct Tools once
We were running the constructor for each Tool object once to run the
`install()` method, and then again for each "init" and "start"
operation.

Now the Tool object constructor runs once, and all code used to
determine if a tool is installed correctly results in the `install()`
method.

We also take the opportunity to properly assert the expected state each
method of the various Tool classes expect to encounter.