Use debug-level environment variables
The remote Tool Meister has a defined environment variable to override
the default "INFO" debug level; however it was not propagated to the
remote over ssh. This adds a command line parameter to directly pass the
value to the remote.

The same logic was replicated in the Tool Data Sink; although "create"
orchestration always forks this on the same host, this allows a simple
and consistent command line hook to enable debug log messages when a
client launches either command directly via "existing" orchestration.