Ensure `${pbench_run}` directory exists
Fixes issue #2665.

We have to unconditionally create the `${pbench_run}` directory in
`agent/base` (and `lib/pbench/agent/base.py` to maintain parallel
behavior) because too many scripts depend on its existence for their
correct function.

In particular, there are bench-scripts which try to report errors but
if the `${pbench_run}` directory does not exist already, the
`${pbench_log}` file location won't exist and the actual error is lost.

Rather than trying to fix the myriad number of places with this
assumption, it is best to just create the directory to meet the assumed
state.