Move metadata creation to linpack.sh
to ensure we are not using outdated files let's move the linpack
metadata creation to the linpack.sh (part of it was already there)
and move the files instead of copying them. The main benefit here
should be not relying on files from previous iteration which might not
have been updated.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>

Add working unit tests for `--clients`

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>