Drop unused `mock_tar_failure`
The `mock_tar_failure` is never invoked when `._send_directory()` is
tested because `._create_tar()` is mocked out.  Drop its use from
`test_tar_create_failure`.

For clarity, rename `fake_tar` to `fake_create_tar` so that it is a
bit more clear that it is used for mocking out `._create_tar` and not
for mocking out the `tar` command somehow.