Remove `mock_tar*` fixtures
The `mock_tar*` fixtures were only used by individual tests. We move
their functionality to the test that used it and then added a new
fixture for creating the `ToolMeister` instance for the test.

We also moved to using `@staticmethods` in `TestCreateTar` since the
`self` parameter of the test object was never used.