Add `relaxed` mode to `check_install_rpm`
Sometimes people might need to disable the `check_install_rpm`
requirement to only use RPM-provided benchmarks, allowing them to use
benchmarks of a different version.

We add an environment variable, `PBENCH_RPM_REQUIREMENT_MODE`, which can
be set to `strict` (the default) or `relaxed`.  In `strict` mode, if the
benchmark RPM requirement is not met, an error message is emitted, and
the execution of the benchmark wrapper script is terminated.  In
`relaxed` mode, a warning message is emitted, and the benchmark script
is allowed to continue.

The warning message also directs the user to run the benchmark with
`strict` mode if problems are encountered before reporting an error.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>
Signed-off-by: Peter Portante <peter.portante@redhat.com>
Co-authored-by: Peter Portante <peter.portante@redhat.com>