Re-try `tar` command if errors encountered
Fixes PR #2692.

When tar'ing up a directory (tool or sysinfo data) to send remotely we
can encounter errors in the operation of `tar` itself.  We add the
`--warning=none` switch to try to force `tar` to make a best-effort at
creating the tar ball if the first attempt fails.  We capture what that
failure was, and record it so that partial results will be explained.

Unit tests were added for the new tar ball creation method, and for the
calling function.

Co-authored-by: Peter Portante <peter.portante@redhat.com>