Refactor to tool meister start tool checks working
The first major change here is to reduce the changes to the TDS to only
be the recording of the TM data in Redis.  The policy about what to do
with tool check failures is left to the driver, since another system
could allow runs with what is still working.

Second, since we are no longer sharing code between modules because the
TDS has been removed from policy decisions. The error handling and
messaging needs to be the way `pbench-tool-meister-start` as well.

Finally, the test is now working cleanly with a fix to correct how the
terminate message is sent during cleanup (it has been broken for a very
long time).

Rebase and Remove unnecessary Code