Port various features from pbench-run-benchmark to pbench-linpack
linpack could be executed via the "pbench-run-benchmark" where it had
"linpack-postprocess-cdm" postprocess script tailored for it. The same
same script is reused for pbench-linpack. To inherit the file structure
we copy the result file to the expected directory and then simply run
the commands.

When executing linpack via pbench-run-benchmark one had the opportunity
to specify various attributes, while the original pbench-linpack relied
on the prepared linpack.dat file. We utilize the already existing code
to allow such customization in pbench-linpack as well.

This is a go-for-broke refactoring which adjusts how `pbench-linpack` is
written, drops the use of `check_install_rpm` in favor of a default
install directory that can be overridden, adding iterations to the
`metdata.log` file, and adds a set of unit tests covering its basic
behaviors.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>
Co-authored-by: Peter Portante <peter.portante@redhat.com>