Remove `pbench-sysinfo-dump` reliance on `base`
There are two environment variables provided by the `agent/base` script
which `pbench-sysinfo-dump` was relying on to have in place.  One was
explicitly set by the Tool Meister in the environment used to run the
script, `pbench_install_dir`, the other was inherited,
`_pbench_full_hostname`.

We change `pbench-sysinfo-dump` to use `sysinfo_install_dir` and
`sysinfo_full_hostname` instead, with the Tool Meister setting those
environment variables explicitly.

We also take the opportunity to remove a useless `_PBENCH_UNIT_TESTS`
reference which is not used due to the existence of the mock,
`agent/util-scripts/test-bin/pbench-sysinfo-dump`.