PBENCH-199 nodeJS Endpoints mirror
This includes two pieces:

1. We introduce an early call to the server's /api/v1/endpoints API
   to identify the full URL for each functional API endpoint;
2. We introduce a nodeJS "mirror" that will run under local npm
   development mode to gather the endpoint configuration of a
   remote Pbench server.

For now, we demonstrate that we successfully fetched the endpoint
definitions by adding the server's identification string to the
"Pbench Dashboard" text. Essentially, to call `datasets/list` you'd
now reference the API map as `window.endpoints.api.datasets_list`.

The idea here is that `.env` defines a `PBENCH_SERVER` environment
variable to point to the actual Pbench server when running the dashboard
in local dev mode. The `index.html` was "hacked" to read the server
endpoints definition from `/api/v1/endpoints`. In a production mode,
where the dashboard `index.html` and Javascript are loaded from the
server, this will resolve directly to the server endpoints definitions.
In dev mode (`npm run dev`) we'll start up a small nodeJS express
"mirror" that simply listens on `localhost:3001/api/v1/endpoints`,
makes the call to the endpoints API on the configured Pbench server,
and returns the results. The `index.html` stores this data on the
Javascript window object as `window.endpoints` for use anywhere in
the Pbench dashboard application. (This was just the way it worked
out in the old dashboard, from which I lifted the `index.html` code; if
you have a better idea, let's do it.)

This adds a `"proxy"` definition to the `package.json`, which tells npm
to proxy access to port 3001 (the express app) through the primary
dashboard port (usually 3000) so the app will respond to the request
on `window.origin`. This also adds `nvm` as a dependency, which (in
the process of floundering to run this with apparently a very old npm)
I found to ensure I was getting the right npm code.