Use new `driver/linpack` script
Instead of a complicated method of writing a script file using `echo`
statements, we create a driver script for linpack which takes care of
transform a set of command line parameters into the form LINPACK
expects.

We also provide a new `linpack-wait` to hide the method of determining
when LINPACK completes.