Not done yet -- max header length 254
The header and sub-header lines in a `linpack.input` file can only be
at most 254 characters (plus a new-line character).  So the headers
are truncated unconditionally.

Since the idea was to place the final command line used in the header,
we now write the final command to a file, `linpack.cmd`, in the output
directory in case the header is truncated.