Fix unit testing for email_validator
The email_validator package just upgraded to 1.2, which implements
stricter validation and in particular now rejects the "@example.com"
addresses used in our unit tests.

This adds a mock to replace use of the real email_validator package in
unit testing so that we're in control (as we should be for all external
dependencies).