Improved `linpack` benchmark wrapper
Instead of a complicated method of writing a script file using `echo`
statements, we create a driver script for linpack which takes care of
transforming a set of command line parameters into the form LINPACK
expects.

We also provide a new `linpack-wait` to hide the method of determining
when LINPACK completes.  The RPM packaging has been updated to deliver
these new files.

These changes remove the need for a per-iteration generated script,
which eliminates the need to copy extra files for each iteration and/or
sample.  Each execution is now per-sample, which allows us to drop the
logic of sharing setup between samples of a given iteration.

Since the `pbench-run-benchmark` script is deprecated, and this commit
completes the implementation of `pbench-linpack` to replace its
functionality, we take a number of steps to:

 * Remove `linpack` support from `pbench-run-benchmark`
 * Rename `linpack-postprocess` to `...-sample`
 * Rename `linpack-postprocess-cdm` to `linpack-postprocess`

We also:

 * Correct `pbench-linpack -C` short option processing

   The `-C` short option was not being handled at all.  We also fix up a
   bit of the formatting for option processing while we're at it.

 * Max header length 254 for `linpack.input` file

   The header and sub-header lines in a `linpack.input` file can only be
   at most 254 characters (plus a new-line character).  So the headers
   are truncated unconditionally.

   Since the idea was to place the final command line used in the header,
   we now write the final command to a file, `linpack.cmd`, in the output
   directory in case the header is truncated.

 * Remove use of `mkdir -p` where possible

 * Start using "last sample" parameter

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>